Separate search bar and write a manual to component.
